Convert page components to TypeScript
